+++ libexec/scoop-search.ps1
@@ -1,28 +1,32 @@
 # Usage: scoop search [<OPTIONS>] [<QUERY>]
 # Summary: Search for applications, which are available for installation.
-#
 # Help: When <QUERY> parameter is not provided, all available applications will be shown.
-# <QUERY> Parameter could be regular expression.
+# <QUERY> parameter could be regular expression.
 #
 # Options:
 #   -h, --help      Show help for this command.
 #   -r, --remote    Force remote search in known buckets using Github API.
-#                       Remote search does not utilize advanced search methods (descriptions, binary, shortcuts, ... matching).
-#                       It only uses manifest name to search.
+#                   Remote search does not utilize advanced search methods (descriptions, binary, shortcuts, ... matching).
+#                   It only uses manifest name to search.
 #   -a, --api       Use shovel.ash258.com API backend for search.
-#                       Regular expression is not supported.
-#                       Currently in experimental phase with limited feature-set. Search is done on manifest names only.
+#                   Regular expression is not supported.
+#                   Currently in experimental phase with limited feature-set. Search is done on manifest names only.
 
-'getopt', 'buckets', 'Helpers', 'Search' | ForEach-Object {
+'core', 'Helpers', 'getopt', 'buckets', 'Search' | ForEach-Object {
     . (Join-Path $PSScriptRoot "..\lib\$_.ps1")
 }
 
 Reset-Alias
 
-$opt, $Query, $err = getopt $args 'ra' 'remote', 'api'
-if ($err) { Stop-ScoopExecution -Message "scoop search: $err" -ExitCode 2 }
-$Remote = $opt.r -or $opt.remote
-$Api = $opt.a -or $opt.api
+$ExitCode = 0
+$LocalResults = @()
+$Options, $Query, $_err = getopt $args 'ra' 'remote', 'api'
+
+if ($_err) { Stop-ScoopExecution -Message "scoop search: $_err" -ExitCode 2 }
+
+$Remote = $Options.r -or $Options.remote
+$Api = $Options.a -or $Options.api
+
 if ($Query) {
     try {
         $Query = New-Object System.Text.RegularExpressions.Regex $Query, 'IgnoreCase'
@@ -33,16 +37,13 @@ if ($Query) {
     $Query = $null
 }
 
-$exitCode = 0
-
-Write-Host 'Searching in local buckets ...'
-$localResults = @()
+Write-UserMessage -Message 'Searching in local buckets...'
 
 foreach ($bucket in (Get-LocalBucket)) {
     $result = Search-LocalBucket -Bucket $bucket -Query $Query
     if (!$result) { continue }
 
-    $localResults += $result
+    $LocalResults += $result
     foreach ($res in $result) {
         Write-Host "$bucket" -ForegroundColor 'Yellow' -NoNewline
         Write-Host '/' -NoNewline
@@ -71,17 +72,17 @@ foreach ($bucket in (Get-LocalBucket)) {
     }
 }
 
-if (!$localResults) { Write-UserMessage -Message 'No matches in local buckets found' }
-if (!$Api -and (!$localResults -or $Remote)) {
+if (!$LocalResults) { Write-UserMessage -Message 'No matches in local buckets found' }
+if (!$Api -and (!$LocalResults -or $Remote)) {
     if (!(Test-GithubApiRateLimitBreached)) {
-        Write-Host 'Searching in remote buckets ...'
+        Write-UserMessage -Message 'Searching in remote buckets ...'
         $remoteResults = Search-AllRemote -Query $Query
 
         if ($remoteResults) {
-            Write-Host "`nResults from other known buckets:`n"
+            Write-UserMessage -Message "`nResults from other known buckets:`n"
             foreach ($r in $remoteResults) {
-                Write-Host "'$($r.bucket)' bucket (Run 'scoop bucket add $($r.bucket)'):"
-                $r.results | ForEach-Object { "    $_" }
+                Write-UserMessage -Message "'$($r.bucket)' bucket (Run 'scoop bucket add $($r.bucket)'):"
+                $r.results | ForEach-Object { Write-UserMessage -Message "    $_" }
             }
         } else {
             Stop-ScoopExecution 'No matches in remote buckets found'
@@ -101,4 +102,4 @@ if ($Api) {
     }
 }
 
-exit $exitCode
+exit $ExitCode