+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 # Changelog
 
+## [0.6.5](https://github.com/Ash258/Scoop-Core/milestone/5)
+
+- **scoop-cat**: Add `-f`, `--format` options
+- Adopt new resolve function for parameter passing
+    - **scoop-home**
+    - **scoop-cat**
+    - **scoop-download**
+
 ## [0.6](https://github.com/Ash258/Scoop-Core/milestone/4)
+
+### 0.6-pre4
+
+- **Checkver**: Stabilize substitutions resolve
+    - i.e: `$urlNoExt` sometimes was faulty resolved as `${url}NoExt`
+- Add additional debug for jsonpath/xpath evaluation
+- **scoop-bucket**: Fix edge case when there are no buckets added
 
 ### 0.6-pre3
 
@@ -76,7 +91,7 @@
     - Fix `installer.file` exit code from ps1 scripts
     - Fix `installer.keep` inconsitency between powershell scripts and executables
 - **scoop-install**: Fix installlation of different/older versions
-- **scoop-info**: Respect `NO_JUNCTION` config
+- **scoop-info**: Respect `NO_JUNCTIONS` config
 - Add changelog to repository
 - **Autoupdate**: Initial preparations for array support
 - **Manifest**:
@@ -98,7 +113,7 @@
 - **Update**: Ignore merge commits in update log
 - `scoop --version` reports PowerShell version
 - **Depends**: Correctly detect globally installed dependencies while resolving script dependencies
-- **Buckets**:
+- **scoop-bucket**:
     - Indicate successful bucket removal
     - Indicate inability of bucket removal